Split out fork resume tests
